feat(web): show diagnostics when expanding session rows
X-Anneal-Run: cmtrfpzw005gfdbrsrkx38tqj
X-Anneal-Step: 2: Implementation
X-Anneal-Provider: codex
